use nextSibling instead of nextElementSibling because of ShadyDOM quirks
Sometimes ShadyDOM returns an invalid nextElementSibling which is actually in the shadow dom and thereby not a real sibling for the marker element.
Then trying to insert before the sibling causes an exception. This approach works around the issue by relying on a text node right after the marker element.